requirements.txt
flask
flask-cors
pymupdf
PyPDF2==3.0.1
requests
pillow
gunicorn
